requirements.txt
Pyrebase4==4.7.1
flask==3.0.0